Refactor AnkiConnectClient::call
Return type is now generic rather than a serde_json::Value. This means
that outside of call() and the private request struct it uses, there is
no direct dependence on either reqwest or serde_json. As a result,
errors are surface at the call() level rather than than the particular
API call wrapper level; these wrappers are now one- or two-liners,
depening on whether or not they need non-null params.